Ускорил инит

+++ src/net/minecraft/client/gui/settings/TabScreen.java
@@ -2,7 +2,7 @@ package net.minecraft.client.gui.settings;
 
 public interface TabScreen {
 
-	void render(float mx, float my, float ticks, int width);
+	void render(int mx, int my, float ticks, int width);
 	void keyboard(int keycode, char c);
 
 	default void mouseDown(int mouseX, int mouseY, int mouseButton) {}
